Support SPM test target
