+++ Package.swift
@@ -16,7 +16,21 @@ let package = Package(
             name: "AcknowList",
             path: "Sources",
             resources: [.process("Resources")]
-        )
+        ),
+        .testTarget(
+            name: "AcknowListTests",
+            dependencies: ["AcknowList"],
+            exclude: ["Info.plist", "Resources"],
+            resources: [
+                .copy("Pods-acknowledgements-RegexTesting.plist"),
+                .copy("Pods-acknowledgements.plist"),
+                .copy("RegexTesting-GroundTruth-Charts.txt"),
+                .copy("RegexTesting-GroundTruth-TYPFontAwesome.txt"),
+                .copy("Pods-acknowledgements-multi.plist"),
+                .copy("RegexTesting-GroundTruth-Alamofire.txt"),
+                .copy("RegexTesting-GroundTruth-TPKeyboardAvoiding.txt"),
+                .copy("RegexTesting-GroundTruth-pop.txt")]
+        ),
     ],
     swiftLanguageVersions: [.v5]
 )
