+++ Tests/AcknowListTests/Resources/RegexTesting-GroundTruth-Alamofire.txt
