+++ Tests/AcknowListTests/Resources/RegexTesting-GroundTruth-Charts.txt
