+++ Tests/AcknowListTests/Resources/RegexTesting-GroundTruth-TPKeyboardAvoiding.txt
