+++ Tests/AcknowListTests/Resources/RegexTesting-GroundTruth-TYPFontAwesome.txt
